+++ pkg/initialize/middeware_test.go
@@ -1,0 +1,12 @@
+package initialize
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestInitMiddleware(t *testing.T) {
+	err := Middleware()
+	assert.NoError(t, err)
+}